progress bar

+++ core/GeometricTools/layerHandler.py
@@ -192,32 +192,50 @@ class LayerHandler(QObject):
         for lyr in lyrList:
             self.updateOriginalLayerFromUnifiedLayer(lyr, unifiedLyr)
     
-    def buildInputDict(self, inpytLyr, pk = None):
+    def buildInputDict(self, inpytLyr, pk = None, feedback = None, progressDelta = 100):
         """
         Maps inputLyr into a dict with its attributes.
         """
         inputDict = dict()
         request = QgsFeatureRequest().setFlags(QgsFeatureRequest.NoGeometry)
-        for feature in inpytLyr.getFeatures(request):
+        currentProgress = feedback.progress() if feedback else None
+        localTotal = progressDelta/inpytLyr.featureCount() if inpytLyr.featureCount() else 0
+        for current, feature in enumerate(inpytLyr.getFeatures(request)):
+            if feedback:
+                if feedback.isCanceled():
+                    break            
             key = feature[pk] if pk else feature.id()
             inputDict[key] = dict()
             inputDict[key]['featList'] = []
             inputDict[key]['featWithoutGeom'] = feature
+            if feedback:
+                feedback.setProgress(currentProgress + int(localTotal*current))
         return inputDict
     
-    def populateInputDictFeatList(self, lyr, inputDict, pk = None, request = None):
+    def populateInputDictFeatList(self, lyr, inputDict, pk = None, request = None, feedback = None, progressDelta = 100):
         iterator = lyr.getFeatures(request) if request else lyr.getFeatures()
-        for feat in iterator:
+        currentProgress = feedback.progress() if feedback else None
+        localTotal = progressDelta/lyr.featureCount() if lyr.featureCount() else 0
+        for current, feat in enumerate(iterator):
+            if feedback:
+                if feedback.isCanceled():
+                    break  
             fid = feat[pk] if pk else feat.id()
             if fid in inputDict:
                 inputDict[fid]['featList'].append(feat)
+            if feedback:
+                feedback.setProgress(currentProgress + int(localTotal*current))
     
-    def updateOriginalLayer(self, originalLayer, resultLayer, field=None, feedback = None):
-        inputDict = self.buildInputDict(originalLayer, pk = field)
-        self.populateInputDictFeatList(resultLayer, inputDict, pk=field)
+    def updateOriginalLayer(self, originalLayer, resultLayer, field=None, feedback = None, progressDelta = 100, keepFeatures = False):
+        #1- build inputDict structure to store the original state of the layer
+        inputDict = self.buildInputDict(originalLayer, pk = field, feedback = feedback, progressDelta = progressDelta/5) 
+        #2- populate the inputDict with the features from the resultLayer
+        self.populateInputDictFeatList(resultLayer, inputDict, pk=field, feedback = feedback, progressDelta = progressDelta/5)
+        #3- get information from originalLayer and resultLayer
         parameterDict = self.getDestinationParameters(originalLayer)
         coordinateTransformer = self.getCoordinateTransformer(resultLayer, originalLayer)
-        self.updateOriginalLayerFeatures(originalLayer, inputDict, parameterDict = parameterDict, coordinateTransformer = coordinateTransformer)
+        #4- run update original layer
+        self.updateOriginalLayerFeatures(originalLayer, inputDict, parameterDict = parameterDict, coordinateTransformer = coordinateTransformer, keepFeatures = keepFeatures, feedback = feedback, progressDelta = 3*progressDelta/5)
 
     def updateOriginalLayerFromUnifiedLayer(self, lyr, unifiedLyr):
         inputDict = self.buildInputDict(lyr)
@@ -230,14 +248,19 @@ class LayerHandler(QObject):
         coordinateTransformer = self.getCoordinateTransformer(unifiedLyr, layer)
         self.updateOriginalLayerFeatures(lyr, inputDict, parameterDict = parameterDict, coordinateTransformer = coordinateTransformer)
     
-    def updateOriginalLayerFeatures(self, lyr, inputDict, parameterDict = {}, coordinateTransformer = None):
+    def updateOriginalLayerFeatures(self, lyr, inputDict, parameterDict = {}, coordinateTransformer = None, keepFeatures = False, feedback = None, progressDelta = 100):
         """
         Updates lyr using inputDict
         """
         idsToRemove, featuresToAdd, idsToRemove = [], [], []
         lyr.startEditing()
         lyr.beginEditCommand('Updating layer {0}'.format(lyr.name()))
-        for id in inputDict:
+        currentProgress = feedback.progress() if feedback else None
+        localTotal = progressDelta/len(inputDict)
+        for current, id in enumerate(inputDict):
+            if feedback:
+                if feedback.isCanceled():
+                    break
             outFeats = inputDict[id]['featList']
             if len(outFeats) == 0 and id not in idsToRemove: #no output, must delete feature
                 idsToRemove.append(id)
@@ -252,6 +275,9 @@ class LayerHandler(QObject):
                     lyr.changeGeometry(id, geomToUpdate) #faster according to the api
                 featuresToAdd += addedFeatures
                 idsToRemove += [id] if deleteId else []
+            if feedback:
+                feedback.setProgress(currentProgress + int(localTotal*current))
         lyr.addFeatures(featuresToAdd)
-        lyr.deleteFeatures(idsToRemove)
+        if not keepFeatures:
+            lyr.deleteFeatures(idsToRemove)
         lyr.endEditCommand()
